Remove stray implicit conversion to create Durations. Move Await object to its own file.
